feat: reorganize the files into directories
