+++ client/app/src/main/AndroidManifest.xml
@@ -9,10 +9,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".FilterFragmentActivity"></activity>
-        <activity android:name=".MyBooksActivity" />
-        <activity android:name=".RequestActivity"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".Activities.FilterFragmentActivity"></activity>
+        <activity android:name=".Activities.MyBooksActivity" />
+        <activity android:name=".Activities.RequestActivity"></activity>
+        <activity android:name=".Activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -20,8 +20,8 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MyProfileActivity"
-            android:parentActivityName=".MainActivity" />
+            android:name=".Activities.MyProfileActivity"
+            android:parentActivityName=".Activities.MainActivity" />
     </application>
 
 </manifest>
