+++ client/app/src/main/java/com/example/criengine/Activities/UserProfileActivity.java
@@ -1,0 +1,18 @@
+package com.example.criengine.Activities;
+
+import android.os.Bundle;
+
+import com.example.criengine.Activities.ProfileActivity;
+import com.example.criengine.R;
+
+public class UserProfileActivity extends ProfileActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        setContentView(R.layout.activity_user_profile);
+        super.onCreate(savedInstanceState);
+
+        // new UI components
+
+    }
+}
