+++ client/app/src/main/java/com/example/criengine/Objects/Profile.java
@@ -1,4 +1,4 @@
-package com.example.criengine;
+package com.example.criengine.Objects;
 
 import java.util.ArrayList;
