--- client/app/src/main/java/com/example/criengine/UserProfileActivity.java
@@ -1,14 +1,0 @@
-package com.example.criengine;
-
-import android.os.Bundle;
-public class UserProfileActivity extends ProfileActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        setContentView(R.layout.activity_user_profile);
-        super.onCreate(savedInstanceState);
-
-        // new UI components
-
-    }
-}
