+++ client/app/src/main/res/layout/activity_my_books.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyBooksActivity">
+    tools:context=".Activities.MyBooksActivity">
 
     <TextView
         android:id="@+id/titleTextView"
